Moved PKCE Code Challenge Method to more prominent section
fixes: #30227
