+++ js/apps/admin-ui/test/clients/details.ts
@@ -1,0 +1,14 @@
+import { Page } from "@playwright/test";
+import { selectItem, assertSelectValue } from "../utils/form";
+
+function getKeyForCodeExchangeInput(page: Page) {
+  return page.locator("#keyForCodeExchange");
+}
+
+export async function selectKeyForCodeExchangeInput(page: Page, value: string) {
+  await selectItem(page, getKeyForCodeExchangeInput(page), value);
+}
+
+export async function assertKeyForCodeExchangeInput(page: Page, value: string) {
+  await assertSelectValue(getKeyForCodeExchangeInput(page), value);
+}
